agent/eerly_studio/requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-dotenv
requests
langfuse
